refactor: modified the card handler and service to improve encapsulation
